IntensityMetrics

IntensityMetrics.Excitation

IntensityMetrics.Medium

IntensityMetrics.intensity

IntensityMetrics.intensity_sppa

IntensityMetrics.intensity_spta

IntensityMetrics.mechanical_index

IntensityMetrics.peak_np

IntensityMetrics.peak_pp

IntensityMetrics.peak_ptp

IntensityMetrics.Excitation — Type

struct Excitation pulseduration: On-time / duration of the excitation pulse in seconds. dutycycle: [0, 1] f0: Center frequency in Hz end

IntensityMetrics.Medium — Type

struct Medium ρ: Density in kg/m³ c: Speed of sound in m/s end

IntensityMetrics.intensity — Method

intensity(p, M::Medium)
p is expected to be a single number in units of pascal.
Output unit is W/m².

IntensityMetrics.intensity_sppa — Method

intensity_sppa(p, M::Medium, E::Excitation)
Spatial Peak, Pulse Averaged Intensity.
p is expected to be a vector/time series containing only the measured pressure from the excitation pulse.
Output unit is W/cm².

IntensityMetrics.intensity_spta — Method

intensity_spta(p, M::Medium, E::Excitation)
Spatial Peak, Time Averaged Intensity.
p is expected to be a vector/time series containing only the measured pressure from the excitation pulse.
Output unit is W/cm².

IntensityMetrics.mechanical_index — Method

mechanical_index(p, E::Excitation)
p is expected to be a vector/time series containing the measured pressure from the excitation pulse.
Output is dimensionless.

IntensityMetrics.peak_np — Method

peak_np(p)
Peak Negative Pressure.

IntensityMetrics.peak_pp — Method

peak_pp(p)
Peak Positive Pressure.

IntensityMetrics.peak_ptp — Method

peak_ptp(p)
peak pressure measured in peak-to-peak.