IntensityMetrics

IntensityMetrics.Excitation

IntensityMetrics.Medium

IntensityMetrics.Isppa

IntensityMetrics.Ispta

IntensityMetrics.MI

IntensityMetrics.intensity

IntensityMetrics.Excitation — Type

struct Excitation pulseduration: On-time / duration of the excitation pulse in seconds. dutycycle: [0, 1] f0: Center frequency in Hz end

IntensityMetrics.Medium — Type

struct Medium ρ: Density in kg/m³ c: Speed of sound in m/s end

IntensityMetrics.Isppa — Method

Isppa(p, M::Medium, E::Excitation)
Spatial Peak, Pulse Averaged Intensity.
p is expected to be a vector/time series containing only the measured pressure from the excitation pulse.
Output unit is W/cm².

IntensityMetrics.Ispta — Method

Ispta(p, M::Medium, E::Excitation)
Spatial Peak, Time Averaged Intensity.
p is expected to be a vector/time series containing only the measured pressure from the excitation pulse.
Output unit is W/cm².

IntensityMetrics.MI — Method

MI(p, E::Excitation)
Mechanical Index.
p is expected to be a vector/time series containing the measured pressure from the excitation pulse.
Output is dimensionless.

IntensityMetrics.intensity — Method

intensity(p, M::Medium)
p is expected to be a single number in units of pascal.
Output unit is W/m².